refactor(ui): SearchField 공통 스타일 분리 및 종속성 제거
- SearchField.styles.ts 도입으로 페이지 전용 스타일 의존 제거
- SearchField가 공통 스타일을 직접 참조하도록 변경
- 검색 입력 UI의 재사용성/일관성 향상
