+++ frontend/src/components/common/SearchField/SearchField.styles.ts
